+++ usecase/src/main/kotlin/self/project/polydator/categorization/gateways/PolygonResolver.kt
@@ -3,5 +3,5 @@ package self.project.polydator.categorization.gateways
 import self.project.polydator.Polygon
 
 interface PolygonResolver {
-    fun resolveShape(sides: List<Float>): Polygon
+    fun resolveShape(sides: List<Int>): Polygon
 }
